+++ tests/e2e/adminAccountManagement.spec.ts
@@ -89,6 +89,7 @@ describe("Account creation, edition and deletion", () => {
         cy.get("input[id='userName']").should("exist");
         cy.get("input[id='email']").should("exist");
         cy.get("input[id='password']").should("exist");
+        cy.get("input[id='phoneNumber']").should("exist");
         cy.get("input[id='firstName']").should("exist");
         cy.get("input[id='lastName']").should("exist");
         cy.get("select[id='day']").should("exist");
@@ -114,10 +115,7 @@ describe("Account creation, edition and deletion", () => {
 
         // student
         cy.get("input[type='radio']").eq(2).click();
-        cy.get("input[id='immatriculationStatus']").should("exist");
         cy.get("input[id='matriculationId']").should("exist");
-        cy.get("input[id='semesterCount']").should("exist");
-        cy.get("select[id='fieldsOfStudy-1']").should("exist");
     });
 
     it("Show unsaved changes modal", () => {
@@ -177,21 +175,17 @@ describe("Account creation, edition and deletion", () => {
         cy.get("button").contains("Create Account").should("be.enabled");
         cy.get("button").contains("Create Account").click();
 
-        cy.get("input[id='email']").siblings().get("p").invoke("hasClass", "error-message");
+        cy.get("input[id='email']").siblings().get("p").should("have.class", "error-message");
 
         cy.get("input[id='email']").clear().type("valid@valid.de");
-    });
-
-    it("Can enter immatriculationStatus", () => {
-        cy.get("input[id='immatriculationStatus']").type("IN");
     });
 
     it("Can enter matriculationId", () => {
         cy.get("input[id='matriculationId']").type("1234567");
     });
 
-    it("Can enter semesterCount", () => {
-        cy.get("input[id='semesterCount']").clear().type("1");
+    it("Can enter phoneNumber", () => {
+        cy.get("input[id='phoneNumber']").type("+49 123456789");
     });
 
     it("Create account works", () => {
@@ -233,7 +227,7 @@ describe("Account creation, edition and deletion", () => {
         cy.get("button").contains("Save Changes").should("be.enabled");
         cy.get("button").contains("Save Changes").click();
 
-        cy.get("input[id='email']").siblings().get("p").invoke("hasClass", "error-message");
+        cy.get("input[id='email']").siblings().get("p").should("have.class", "error-message");
 
         cy.get("input[id='email']").clear().type("valid@valid.de");
     });
@@ -241,16 +235,6 @@ describe("Account creation, edition and deletion", () => {
     it("Country enum is filled", () => {
         cy.get('select[id="country"]').select("Germany");
         cy.get('select[id="country"]').select("United States");
-    });
-
-    it("Select fields of study is working properly", () => {
-        cy.get("select[id='fieldsOfStudy-1']").select("Economics");
-        cy.get("select[id='fieldsOfStudy-2']").select("Physics");
-        cy.get("select[id='fieldsOfStudy-3']").find("option").contains("Computer Science").should("exist");
-        cy.get("select[id='fieldsOfStudy-3']").find("option").contains("Physics").should("not.exist");
-        cy.get("select[id='fieldsOfStudy-3']").select("Education");
-        cy.get("button[id='removeFieldOfStudy-2']").click();
-        cy.get("select[id='fieldsOfStudy-3']").select("Computer Science");
     });
 
     it("Can change name", () => {
@@ -278,6 +262,10 @@ describe("Account creation, edition and deletion", () => {
     it("Can edit password", () => {
         cy.get("input[id='password']").should("exist");
         cy.get("input[id='password']").type("test-password-cypress");
+    });
+
+    it("Can enter phoneNumber", () => {
+        cy.get("input[id='phoneNumber']").type("+49 123456789");
     });
 
     it("Country enum is filled", () => {
@@ -401,6 +389,10 @@ describe("Account creation, edition and deletion", () => {
     it("Can edit password", () => {
         cy.get("input[id='password']").should("exist");
         cy.get("input[id='password']").type("test-password-cypress");
+    });
+
+    it("Can enter phoneNumber", () => {
+        cy.get("input[id='phoneNumber']").type("+49 123456789");
     });
 
     it("Country enum is filled", () => {