+++ tests/e2e/lecturerCourseManagement.spec.ts
@@ -25,8 +25,8 @@ describe("Course creation, edition and deletion", () => {
     it("Navigate to course list", () => {
         cy.get("div[id='menu_courses']").children().eq(1).should("not.be.visible");
         cy.get("div[id='menu_courses']").parents().eq(1).trigger("mouseover");
-        cy.get("div[id='menu_courses']").children().eq(1).get("span").contains("All Courses").should("be.visible");
-        cy.get("div[id='menu_courses']").children().eq(1).get("a").contains("All Courses").click();
+        cy.get("div[id='menu_courses']").children().eq(1).get("span").contains("My Courses").should("be.visible");
+        cy.get("div[id='menu_courses']").children().eq(1).get("a").contains("My Courses").click();
         cy.get("div[id='menu_courses']").trigger("mouseleave");
         cy.url().should("contain", "course-management");
     });